+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Use the separate `default export`s from `vtex.checkout-resources`.
+
 ## [0.3.0] - 2019-12-10
 
 ### Added